Data.GenericLens.Internal

Description

The library internals are exposed through this module. Please keep in mind that everything here is subject to change irrespective of the the version numbers.

Documentation

type Indexed t = Indexed' t 0

type family Infer (s :: *) (a' :: *) (b :: *) :: * where ...

Equations

Infer (s a) a' b = ReplaceArgs (s a) (Unify a' b)

Infer s _ _ = s

data PTag

Constructors

PTag

type family P :: Nat -> k -> PTag -> k

type family LookupParam (a :: k) (p :: Nat) :: Maybe Nat where ...

Equations

LookupParam (param (n :: Nat)) m = Nothing

LookupParam (a (_ (m :: Nat))) n = IfEq m n (Just 0) (MaybeAdd (LookupParam a n) 1)

LookupParam (a _) n = MaybeAdd (LookupParam a n) 1

LookupParam a _ = Nothing

type family ArgAt (t :: k) (n :: Nat) :: j where ...

Equations

ArgAt (t a) 0 = a

ArgAt (t a) n = ArgAt t (n - 1)

type family ArgCount (t :: k) :: Nat where ...

Equations

ArgCount (f a) = 1 + ArgCount f

ArgCount a = 0

class UnifyHead (a :: k) (b :: k)

Ensure that the types a and b are both applications of the same constructor. The arguments may be different.

type family HasTotalFieldP (field :: Symbol) f :: Maybe Type where ...

Equations

HasTotalFieldP field (S1 (MetaSel (Just field) _ _ _) (Rec0 t)) = Just t

HasTotalFieldP field (l :*: r) = Alt (HasTotalFieldP field l) (HasTotalFieldP field r)

HasTotalFieldP field (l :+: r) = Both (HasTotalFieldP field l) (HasTotalFieldP field r)

HasTotalFieldP field (S1 _ _) = Nothing

HasTotalFieldP field (C1 _ f) = HasTotalFieldP field f

HasTotalFieldP field (D1 _ f) = HasTotalFieldP field f

HasTotalFieldP field (K1 _ _) = Nothing

HasTotalFieldP field U1 = Nothing

HasTotalFieldP field V1 = Nothing

type family HasTotalTypeP (typ :: Type) f :: Maybe Type where ...

Equations

HasTotalTypeP typ (S1 _ (K1 _ typ)) = Just typ

HasTotalTypeP typ (l :*: r) = Alt (HasTotalTypeP typ l) (HasTotalTypeP typ r)

HasTotalTypeP typ (l :+: r) = Both (HasTotalTypeP typ l) (HasTotalTypeP typ r)

HasTotalTypeP typ (S1 _ _) = Nothing

HasTotalTypeP typ (C1 _ f) = HasTotalTypeP typ f

HasTotalTypeP typ (D1 _ f) = HasTotalTypeP typ f

HasTotalTypeP typ (K1 _ _) = Nothing

HasTotalTypeP typ U1 = Nothing

HasTotalTypeP typ V1 = Nothing

type family HasTotalPositionP (pos :: Nat) f :: Maybe Type where ...

Equations

HasTotalPositionP pos (S1 _ (K1 (Pos pos) t)) = Just t

HasTotalPositionP pos (l :*: r) = Alt (HasTotalPositionP pos l) (HasTotalPositionP pos r)

HasTotalPositionP pos (l :+: r) = Both (HasTotalPositionP pos l) (HasTotalPositionP pos r)

HasTotalPositionP pos (S1 _ _) = Nothing

HasTotalPositionP pos (C1 _ f) = HasTotalPositionP pos f

HasTotalPositionP pos (D1 _ f) = HasTotalPositionP pos f

HasTotalPositionP pos (K1 _ _) = Nothing

HasTotalPositionP pos U1 = Nothing

HasTotalPositionP pos V1 = Nothing

data Pos (p :: Nat)

type family HasPartialTypeP a f :: Bool where ...

Equations

HasPartialTypeP t (l :+: r) = HasPartialTypeP t l || HasPartialTypeP t r

HasPartialTypeP t (C1 m f) = t == GTypes f

HasPartialTypeP t (M1 _ _ f) = HasPartialTypeP t f

HasPartialTypeP t _ = False

type family HasCtorP (ctor :: Symbol) f :: Bool where ...

Equations

HasCtorP ctor (C1 (MetaCons ctor _ _) _) = True

HasCtorP ctor (f :+: g) = HasCtorP ctor f || HasCtorP ctor g

HasCtorP ctor (D1 m f) = HasCtorP ctor f

HasCtorP ctor _ = False

type family GTypes (rep :: Type -> Type) :: [Type] where ...

Equations

GTypes (l :*: r) = GTypes l ++ GTypes r

GTypes (K1 _ a) = '[a]

GTypes (M1 _ m a) = GTypes a

GTypes U1 = '[]

type family CollectTotalType t f :: TypeStat where ...

Equations

CollectTotalType t (C1 (MetaCons ctor _ _) f) = AddToStat ctor (CountType t f) EmptyStat

CollectTotalType t (M1 _ _ r) = CollectTotalType t r

CollectTotalType t (l :+: r) = MergeStat (CollectTotalType t l) (CollectTotalType t r)

type family CollectPartialType t f :: [Symbol] where ...

Equations

CollectPartialType t (l :+: r) = CollectPartialType t l ++ CollectPartialType t r

CollectPartialType t (C1 (MetaCons ctor _ _) f) = If (t == GTypes f) '[ctor] '[]

CollectPartialType t (D1 _ f) = CollectPartialType t f

type family CollectField t f :: TypeStat where ...

Equations

CollectField t (C1 (MetaCons ctor _ _) f) = AddToStat ctor (CountField t f) EmptyStat

CollectField t (M1 _ _ r) = CollectField t r

CollectField t (l :+: r) = MergeStat (CollectField t l) (CollectField t r)

type family CollectFieldsOrdered (r :: * -> *) :: [Symbol] where ...

Equations

CollectFieldsOrdered (l :*: r) = Merge (CollectFieldsOrdered l) (CollectFieldsOrdered r)

CollectFieldsOrdered (S1 (MetaSel (Just name) _ _ _) _) = '[name]

CollectFieldsOrdered (M1 _ m a) = CollectFieldsOrdered a

CollectFieldsOrdered _ = '[]

data TypeStat

Constructors

TypeStat

Fields

_containsNone :: [Symbol]

_containsMultiple :: [Symbol]

_containsOne :: [Symbol]

type family (xs :: [Symbol]) \\ (ys :: [Symbol]) :: [Symbol] where ... infixr 5

Equations

xs \\ '[] = xs

'[] \\ xs = '[]

(x ': xs) \\ (y ': ys) = Sub' (CmpSymbol x y) x y xs ys

type family ShowSymbols (ctors :: [Symbol]) :: ErrorMessage where ...

Equations

ShowSymbols '[] = Text ""

ShowSymbols (c ': cs) = (Text "\8226 " :<>: Text c) :$$: ShowSymbols cs

type Indexed t = Indexed' t 0

type family Infer (s :: *) (a' :: *) (b :: *) :: * where ...

Equations

Infer (s a) a' b = ReplaceArgs (s a) (Unify a' b)

Infer s _ _ = s

data PTag

Constructors

PTag

type family P :: Nat -> k -> PTag -> k

type family LookupParam (a :: k) (p :: Nat) :: Maybe Nat where ...

Equations

LookupParam (param (n :: Nat)) m = Nothing

LookupParam (a (_ (m :: Nat))) n = IfEq m n (Just 0) (MaybeAdd (LookupParam a n) 1)

LookupParam (a _) n = MaybeAdd (LookupParam a n) 1

LookupParam a _ = Nothing

type family ArgAt (t :: k) (n :: Nat) :: j where ...

Equations

ArgAt (t a) 0 = a

ArgAt (t a) n = ArgAt t (n - 1)

type family ArgCount (t :: k) :: Nat where ...

Equations

ArgCount (f a) = 1 + ArgCount f

ArgCount a = 0

class UnifyHead (a :: k) (b :: k)

Ensure that the types a and b are both applications of the same constructor. The arguments may be different.

type family CollectTotalType t f :: TypeStat where ...

Equations

CollectTotalType t (C1 (MetaCons ctor _ _) f) = AddToStat ctor (CountType t f) EmptyStat

CollectTotalType t (M1 _ _ r) = CollectTotalType t r

CollectTotalType t (l :+: r) = MergeStat (CollectTotalType t l) (CollectTotalType t r)

type family CollectPartialType t f :: [Symbol] where ...

Equations

CollectPartialType t (l :+: r) = CollectPartialType t l ++ CollectPartialType t r

CollectPartialType t (C1 (MetaCons ctor _ _) f) = If (t == GTypes f) '[ctor] '[]

CollectPartialType t (D1 _ f) = CollectPartialType t f

type family CollectField t f :: TypeStat where ...

Equations

CollectField t (C1 (MetaCons ctor _ _) f) = AddToStat ctor (CountField t f) EmptyStat

CollectField t (M1 _ _ r) = CollectField t r

CollectField t (l :+: r) = MergeStat (CollectField t l) (CollectField t r)

type family CollectFieldsOrdered (r :: * -> *) :: [Symbol] where ...

Equations

CollectFieldsOrdered (l :*: r) = Merge (CollectFieldsOrdered l) (CollectFieldsOrdered r)

CollectFieldsOrdered (S1 (MetaSel (Just name) _ _ _) _) = '[name]

CollectFieldsOrdered (M1 _ m a) = CollectFieldsOrdered a

CollectFieldsOrdered _ = '[]

data TypeStat

Constructors

TypeStat

Fields

_containsNone :: [Symbol]

_containsMultiple :: [Symbol]

_containsOne :: [Symbol]

type family (xs :: [Symbol]) \\ (ys :: [Symbol]) :: [Symbol] where ... infixr 5

Equations

xs \\ '[] = xs

'[] \\ xs = '[]

(x ': xs) \\ (y ': ys) = Sub' (CmpSymbol x y) x y xs ys

type family HasTotalFieldP (field :: Symbol) f :: Maybe Type where ...

Equations

HasTotalFieldP field (S1 (MetaSel (Just field) _ _ _) (Rec0 t)) = Just t

HasTotalFieldP field (l :*: r) = Alt (HasTotalFieldP field l) (HasTotalFieldP field r)

HasTotalFieldP field (l :+: r) = Both (HasTotalFieldP field l) (HasTotalFieldP field r)

HasTotalFieldP field (S1 _ _) = Nothing

HasTotalFieldP field (C1 _ f) = HasTotalFieldP field f

HasTotalFieldP field (D1 _ f) = HasTotalFieldP field f

HasTotalFieldP field (K1 _ _) = Nothing

HasTotalFieldP field U1 = Nothing

HasTotalFieldP field V1 = Nothing

type family HasTotalTypeP (typ :: Type) f :: Maybe Type where ...

Equations

HasTotalTypeP typ (S1 _ (K1 _ typ)) = Just typ

HasTotalTypeP typ (l :*: r) = Alt (HasTotalTypeP typ l) (HasTotalTypeP typ r)

HasTotalTypeP typ (l :+: r) = Both (HasTotalTypeP typ l) (HasTotalTypeP typ r)

HasTotalTypeP typ (S1 _ _) = Nothing

HasTotalTypeP typ (C1 _ f) = HasTotalTypeP typ f

HasTotalTypeP typ (D1 _ f) = HasTotalTypeP typ f

HasTotalTypeP typ (K1 _ _) = Nothing

HasTotalTypeP typ U1 = Nothing

HasTotalTypeP typ V1 = Nothing

type family HasTotalPositionP (pos :: Nat) f :: Maybe Type where ...

Equations

HasTotalPositionP pos (S1 _ (K1 (Pos pos) t)) = Just t

HasTotalPositionP pos (l :*: r) = Alt (HasTotalPositionP pos l) (HasTotalPositionP pos r)

HasTotalPositionP pos (l :+: r) = Both (HasTotalPositionP pos l) (HasTotalPositionP pos r)

HasTotalPositionP pos (S1 _ _) = Nothing

HasTotalPositionP pos (C1 _ f) = HasTotalPositionP pos f

HasTotalPositionP pos (D1 _ f) = HasTotalPositionP pos f

HasTotalPositionP pos (K1 _ _) = Nothing

HasTotalPositionP pos U1 = Nothing

HasTotalPositionP pos V1 = Nothing

data Pos (p :: Nat)

type family HasPartialTypeP a f :: Bool where ...

Equations

HasPartialTypeP t (l :+: r) = HasPartialTypeP t l || HasPartialTypeP t r

HasPartialTypeP t (C1 m f) = t == GTypes f

HasPartialTypeP t (M1 _ _ f) = HasPartialTypeP t f

HasPartialTypeP t _ = False

type family HasCtorP (ctor :: Symbol) f :: Bool where ...

Equations

HasCtorP ctor (C1 (MetaCons ctor _ _) _) = True

HasCtorP ctor (f :+: g) = HasCtorP ctor f || HasCtorP ctor g

HasCtorP ctor (D1 m f) = HasCtorP ctor f

HasCtorP ctor _ = False

type family GTypes (rep :: Type -> Type) :: [Type] where ...

Equations

GTypes (l :*: r) = GTypes l ++ GTypes r

GTypes (K1 _ a) = '[a]

GTypes (M1 _ m a) = GTypes a

GTypes U1 = '[]

data Void

data Void1 a

data Void2 a b

class GAsConstructor (ctor :: Symbol) s t a b | ctor s -> a, ctor t -> b where

As AsConstructor but over generic representations as defined by GHC.Generics.

Methods

_GCtor :: Prism (s x) (t x) a b

type GAsConstructor' ctor s a = GAsConstructor ctor s s a a

class GAsType (f :: Type -> Type) (as :: [Type]) where

As AsType but over generic representations as defined by GHC.Generics.

Methods

_GTyped :: Prism (f x) (f x) (HList as) (HList as)

class GAsSubtype (subf :: Type -> Type) (supf :: Type -> Type) where

As AsSubtype but over generic representations as defined by GHC.Generics.

Methods

_GSub :: Prism' (supf x) (subf x)

type (<?) x y = Not (y <=? x) infixl 4

type family Size f :: Nat where ...

Equations

Size (l :*: r) = Size l + Size r

Size (l :+: r) = Min (Size l) (Size r)

Size (D1 meta f) = Size f

Size (C1 meta f) = Size f

Size f = 1

type family CRep (a :: Type) :: G where ...

In-order labeling of the generic tree with the field positions

We replace the (K1 R a) nodes with (K1 (Pos n) a), where n is the position of the field in question in the data type. This is convenient, because we can reuse all the existing functions as long as they are polymorphic in the first parameter of K1.

Equations

CRep rep = Fst (Traverse (Rep rep) 1)

class GLens (pred :: Pred) (s :: Type -> Type) (t :: Type -> Type) a b | s pred -> a, t pred -> b where

Methods

glens :: Lens (s x) (t x) a b

type GLens' pred s a = GLens pred s s a a

type TyFun a b = a -> b -> Type

type family Eval (f :: TyFun a b) (x :: a) :: b

class GUpcast (sub :: Type -> Type) (sup :: Type -> Type) where

Methods

gupcast :: sub p -> sup p

class GSmash sub sup where

Methods

gsmash :: sup p -> sub p -> sub p

class GHasConstraints (c :: * -> * -> Constraint) s t where

Methods

gconstraints :: TraversalC c (s x) (t x)

class GHasConstraints' (c :: * -> Constraint) (f :: * -> *) where

Constrained traversal.

Methods

gconstraints' :: forall g x. Applicative g => (forall a. c a => a -> g a) -> f x -> g (f x)

class GIsList (f :: Type -> Type) (g :: Type -> Type) (as :: [Type]) (bs :: [Type]) | f -> as, g -> bs, bs f -> g, as g -> f where

Minimal complete definition

glist

Methods

glist :: Iso (f x) (g x) (HList as) (HList bs)

glistR :: Iso (HList bs) (HList as) (g x) (f x)

class IndexList (i :: Nat) as bs a b | i as -> a, i bs -> b, i as b -> bs, i bs a -> as where

Methods

point :: Lens (HList as) (HList bs) a b

data HList (as :: [Type]) where

Constructors

Nil :: HList '[]

(:>) :: a -> HList as -> HList (a ': as) infixr 5

type family (as :: [k]) ++ (bs :: [k]) :: [k] where ...

Equations

'[] ++ bs = bs

(a ': as) ++ bs = a ': (as ++ bs)

class Elem (as :: [(k, Type)]) (key :: k) (i :: Nat) a | as key -> i a

class ListTuple (tuple :: Type) (as :: [Type]) | as -> tuple where

Minimal complete definition

tupleToList, listToTuple

Associated Types

type ListToTuple as :: Type

Methods

tupled :: Iso' (HList as) tuple

tupleToList :: tuple -> HList as

listToTuple :: HList as -> tuple

type family TupleToList a where ...

Equations

TupleToList () = '[]

TupleToList (a, b) = '[a, b]

TupleToList (a, b, c) = '[a, b, c]

TupleToList (a, b, c, d) = '[a, b, c, d]

TupleToList (a, b, c, d, e) = '[a, b, c, d, e]

TupleToList (a, b, c, d, e, f) = '[a, b, c, d, e, f]

TupleToList (a, b, c, d, e, f, g) = '[a, b, c, d, e, f, g]

TupleToList (a, b, c, d, e, f, g, h) = '[a, b, c, d, e, f, g, h]

TupleToList (a, b, c, d, e, f, g, h, j) = '[a, b, c, d, e, f, g, h, j]

TupleToList (a, b, c, d, e, f, g, h, j, k) = '[a, b, c, d, e, f, g, h, j, k]

TupleToList (a, b, c, d, e, f, g, h, j, k, l) = '[a, b, c, d, e, f, g, h, j, k, l]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m) = '[a, b, c, d, e, f, g, h, j, k, l, m]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n) = '[a, b, c, d, e, f, g, h, j, k, l, m, n]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r]

TupleToList (a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r, s) = '[a, b, c, d, e, f, g, h, j, k, l, m, n, o, p, q, r, s]

TupleToList a = '[a]