Data.Generics.Internal.Profunctor.Lens

Description

Internal lens helpers. Only exported for Haddock

Documentation

type Lens s t a b = forall p. Strong p => p a b -> p s t

type LensLike p s t a b = p a b -> p s t

ravel :: (ALens a b a b -> ALens a b s t) -> Lens s t a b

set :: ((a -> b) -> s -> t) -> (s, b) -> t

Setting

view :: Lens s s a a -> s -> a

withLensPrim :: Lens s t a b -> (forall c. (s -> (c, a)) -> ((c, b) -> t) -> r) -> r

idLens :: ALens a b a b

first :: Lens ((a :*: b) x) ((a' :*: b) x) (a x) (a' x)

Lens focusing on the first element of a product

second :: Lens ((a :*: b) x) ((a :*: b') x) (b x) (b' x)

Lens focusing on the second element of a product

fork :: (a -> b) -> (a -> c) -> a -> (b, c)

swap :: (a, b) -> (b, a)

cross :: (a -> b) -> (c -> d) -> (a, c) -> (b, d)

data Coyoneda f b

Constructors

Coyoneda (a -> b) (f a)

inj :: Functor f => Coyoneda f a -> f a

proj :: Functor f => f a -> Coyoneda f a

newtype Alongside p s t a b

Constructors

Alongside

Fields

getAlongside :: p (s, a) (t, b)

(??) :: Functor f => f (a -> b) -> a -> f b

alongside :: Profunctor p => LensLike (Alongside p s' t') s t a b -> LensLike (Alongside p a b) s' t' a' b' -> LensLike p (s, s') (t, t') (a, a') (b, b')

assoc3L :: Lens ((a, b), c) ((a', b'), c') (a, (b, c)) (a', (b', c'))

stron :: (Either s s', b) -> Either (s, b) (s', b)

choosing :: forall s t a b s' t'. Lens s t a b -> Lens s' t' a b -> Lens (Either s s') (Either t t') a b

lens :: (s -> (c, a)) -> ((c, b) -> t) -> Lens s t a b

data ALens a b s t

Constructors

ALens (s -> (c, a)) ((c, b) -> t)

mLens :: Lens (M1 i c f p) (M1 i c g p) (f p) (g p)

repLens :: (Generic a, Generic b) => Lens a b (Rep a x) (Rep b x)

prodL :: Lens ((a :*: b) x) ((a' :*: b') x) (a x, b x) (a' x, b' x)

prodR :: Lens (a' x, b' x) (a x, b x) ((a' :*: b') x) ((a :*: b) x)

assoc3R :: Lens (a', (b', c')) (a, (b, c)) ((a', b'), c') ((a, b), c)