Data.Generics.Product

Description

Magic product operations using Generics

These classes need not be instantiated manually, as GHC can automatically prove valid instances via Generics. Only the Generic class needs to be derived (see examples).

Lenses

module Data.Generics.Product.Any

module Data.Generics.Product.Fields

module Data.Generics.Product.Positions

module Data.Generics.Product.Subtype

module Data.Generics.Product.Typed

module Data.Generics.Product.HList

Traversals

module Data.Generics.Product.Types

module Data.Generics.Product.Param