Data.Generics.Product.Positions

Description

Derive positional product type getters and setters generically.

Lenses

Running example:

>>> :set -XTypeApplications
>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> :set -XGADTs
>>> :set -XFlexibleContexts
>>> import GHC.Generics
>>> :m +Data.Generics.Internal.VL.Lens
>>> :m +Data.Function
>>> :{
data Human = Human
  { name    :: String
  , age     :: Int
  , address :: String
  }
  deriving (Generic, Show)
human :: Human
human = Human "Tunyasz" 50 "London"
:}

class HasPosition (i :: Nat) s t a b | s i -> a, t i -> b, s i b -> t, t i a -> s where

Records that have a field at a given position.

Methods

position :: Lens s t a b

A lens that focuses on a field at a given position. Compatible with the lens package's Lens type.

>>> human ^. position @1
"Tunyasz"
>>> human & position @3 .~ "Berlin"
Human {name = "Tunyasz", age = 50, address = "Berlin"}

Type errors

>>> human & position @4 .~ "Berlin"
...
... The type Human does not contain a field at position 4
...

class HasPosition' (i :: Nat) s a | s i -> a where

Records that have a field at a given position.

The difference between HasPosition and HasPosition_ is similar to the one between HasField and HasField_. See HasField_.

Methods

position' :: Lens s s a a

class HasPosition_ (i :: Nat) s t a b where

Methods

position_ :: Lens s t a b

class HasPosition0 (i :: Nat) s t a b where

Records that have a field at a given position.

This class gives the minimal constraints needed to define this lens. For common uses, see HasPosition.

Methods

position0 :: Lens s t a b

getPosition :: forall i s a. HasPosition' i s a => s -> a

>>> getPosition @2 human
50

setPosition :: forall i s a. HasPosition' i s a => a -> s -> s

>>> setPosition @2 60 human
Human {name = "Tunyasz", age = 60, address = "London"}