Data.Generics.Product.Subtype

Description

Structural subtype relationships between product types.

Lenses

Running example:

>>> :set -XTypeApplications
>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> :set -XDuplicateRecordFields
>>> import GHC.Generics
>>> :m +Data.Generics.Internal.VL.Lens
>>> :{
data Human = Human
  { name    :: String
  , age     :: Int
  , address :: String
  }
  deriving (Generic, Show)
data Animal = Animal
  { name    :: String
  , age     :: Int
  }
  deriving (Generic, Show)
human :: Human
human = Human "Tunyasz" 50 "London"
:}

class Subtype sup sub where

Structural subtype relationship

sub is a (structural) subtype of sup, if its fields are a subset of those of sup.

Minimal complete definition

super | smash, upcast

Methods

super :: Lens sub sub sup sup

Structural subtype lens. Given a subtype relationship sub :< sup, we can focus on the sub structure of sup.

>>> human ^. super @Animal
Animal {name = "Tunyasz", age = 50}

>>> set (super @Animal) (Animal "dog" 10) human
Human {name = "dog", age = 10, address = "London"}

upcast :: sub -> sup

Cast the more specific subtype to the more general supertype

>>> upcast human :: Animal
Animal {name = "Tunyasz", age = 50}

>>> upcast (upcast human :: Animal) :: Human
...
... The type 'Animal' is not a subtype of 'Human'.
... The following fields are missing from 'Animal':
... address
...

smash :: sup -> sub -> sub

Plug a smaller structure into a larger one

>>> smash (Animal "dog" 10) human
Human {name = "dog", age = 10, address = "London"}