Data.Generics.Sum.Any

Description

Derive a variety of prisms generically.

Prisms

Running example:

>>> :set -XTypeApplications
>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> import GHC.Generics
>>> import Control.Lens
>>> :{
data Animal
  = Dog Dog
  | Cat Name Age
  | Duck Age
  deriving (Generic, Show)
data Dog
  = MkDog
  { name :: Name
  , age  :: Age
  }
  deriving (Generic, Show)
type Name = String
type Age  = Int
dog, cat, duck :: Animal
dog = Dog (MkDog "Shep" 3)
cat = Cat "Mog" 5
duck = Duck 2
:}

class AsAny sel a s | s sel -> a where

Sums that have generic prisms.

Methods

_As :: Prism s s a a

A prism that projects a sum as identified by some selector. Currently supported selectors are constructor names and unique types. Compatible with the lens package's Prism type.

>>> dog ^? _As @"Dog"
Just (MkDog {name = "Shep", age = 3})

>>> dog ^? _As @Dog
Just (MkDog {name = "Shep", age = 3})

>>> dog ^? _As @"Cat"
Nothing

>>> cat ^? _As @(Name, Age)
Just ("Mog",5)

>>> cat ^? _As @"Cat"
Just ("Mog",5)

>>> _As @"Cat" # ("Garfield", 6) :: Animal
Cat "Garfield" 6

>>> duck ^? _As @Age
Just 2