Data.Generics.Sum.Constructors

Description

Derive constructor-name-based prisms generically.

Prisms

Running example:

>>> :set -XTypeApplications
>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> :set -XFlexibleContexts
>>> :set -XTypeFamilies
>>> import GHC.Generics
>>> import Control.Lens
>>> :m +Data.Generics.Product.Fields
>>> :m +Data.Function
>>> :{
data Animal a
  = Dog (Dog a)
  | Cat Name Age
  | Duck Age
  deriving (Generic, Show)
data Dog a
  = MkDog
  { name   :: Name
  , age    :: Age
  , fieldA :: a
  }
  deriving Show
type Name = String
type Age  = Int
dog, cat, duck :: Animal Int
dog = Dog (MkDog "Shep" 3 30)
cat = Cat "Mog" 5
duck = Duck 2
:}

class AsConstructor (ctor :: Symbol) s t a b | ctor s -> a, ctor t -> b where

Sums that have a constructor with a given name.

Methods

_Ctor :: Prism s t a b

A prism that projects a named constructor from a sum. Compatible with the lens package's Prism type.

>>> dog ^? _Ctor @"Dog"
Just (MkDog {name = "Shep", age = 3, fieldA = 30})

>>> dog ^? _Ctor @"Cat"
Nothing

>>> cat ^? _Ctor @"Cat"
Just ("Mog",5)

>>> _Ctor @"Cat" # ("Garfield", 6) :: Animal Int
Cat "Garfield" 6

Type errors

>>> cat ^? _Ctor @"Turtle"
...
...
... The type Animal Int does not contain a constructor named "Turtle"
...

class AsConstructor_ (ctor :: Symbol) s t a b where

Sums that have a constructor with a given name.

The difference between HasConstructor and HasConstructor_ is similar to the one between HasField and HasField_. See HasField_.

Methods

_Ctor_ :: Prism s t a b

class AsConstructor' (ctor :: Symbol) s a | ctor s -> a where

Methods

_Ctor' :: Prism s s a a

class AsConstructor0 (ctor :: Symbol) s t a b where

Sums that have a constructor with a given name.

This class gives the minimal constraints needed to define this prism. For common uses, see HasConstructor.

Methods

_Ctor0 :: Prism s t a b