Data.Generics.Sum.Typed

Description

Derive constructor-field-type-based prisms generically.

Prisms

class AsType a s where

Sums that have a constructor with a field of the given type.

Minimal complete definition

injectTyped, projectTyped | _Typed

Methods

_Typed :: Prism' s a

A prism that projects a constructor uniquely identifiable by the type of its field. Compatible with the lens package's Prism type.

>>> dog ^? _Typed @Dog
Just (MkDog {name = "Shep", age = Age 3})
>>> cat ^? _Typed @(Name, Age)
Just ("Mog",Age 5)
>>> dog ^? _Typed @Age
...
...
... The type Animal contains multiple constructors whose fields are of type Age.
... The choice of constructor is thus ambiguous, could be any of:
... Duck
... Turtle
...

injectTyped :: a -> s

Inject by type.

projectTyped :: s -> Maybe a

Project by type.