Data.Generics.Product.Any

Description

Derive a variety of lenses generically.

Lenses

Running example:

>>> :set -XTypeApplications
>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> import GHC.Generics
>>> :m +Data.Generics.Internal.VL.Lens
>>> :{
data Human = Human
  { name    :: String
  , age     :: Int
  , address :: String
  }
  deriving (Generic, Show)
human :: Human
human = Human "Tunyasz" 50 "London"
:}

class HasAny (sel :: k) s t a b | s sel k -> a where

Methods

the :: Lens s t a b

A lens that focuses on a part of a product as identified by some selector. Currently supported selectors are field names, positions and unique types. Compatible with the lens package's Lens type.

>>> human ^. the @Int
50

>>> human ^. the @"name"
"Tunyasz"

>>> human ^. the @3
"London"