Data.Generics.Product.Fields

Description

Derive record field getters and setters generically.

Lenses

Running example:

>>> :set -XTypeApplications
>>> :set -XDataKinds
>>> :set -XDeriveGeneric
>>> :set -XGADTs
>>> :set -XFlexibleContexts
>>> import GHC.Generics
>>> :m +Data.Generics.Internal.VL.Lens
>>> :m +Data.Function
>>> :{
data Human a
  = Human
    { name    :: String
    , age     :: Int
    , address :: String
    , other   :: a
    }
  | HumanNoAddress
    { name    :: String
    , age     :: Int
    , other   :: a
    }
  deriving (Generic, Show)
human :: Human Bool
human = Human { name = "Tunyasz", age = 50, address = "London", other = False }
:}

class HasField (field :: Symbol) s t a b | s field -> a, t field -> b, s field b -> t, t field a -> s where

Records that have a field with a given name.

Methods

field :: Lens s t a b

A lens that focuses on a field with a given name. Compatible with the lens package's Lens type.

>>> human ^. field @"age"
50

Type changing

>>> :t human
human :: Human Bool

>>> :t human & field @"other" .~ (42 :: Int)
human & field @"other" .~ (42 :: Int) :: Human Int

>>> human & field @"other" .~ 42
Human {name = "Tunyasz", age = 50, address = "London", other = 42}

Type errors

>>> human & field @"weight" .~ 42
...
... The type Human Bool does not contain a field named 'weight'.
...

>>> human & field @"address" .~ ""
...
... Not all constructors of the type Human Bool
... contain a field named 'address'.
... The offending constructors are:
... HumanNoAddress
...

class HasField' (field :: Symbol) s a | s field -> a where

Methods

field' :: Lens s s a a

class HasField_ (field :: Symbol) s t a b where

Records that have a field with a given name.

This is meant to be more general than HasField, but that is not quite the case due to the lack of functional dependencies.

The types s and t must be applications of the same type constructor. In contrast, HasField also requires the parameters of that type constructor to have representational roles.

One use case of HasField_ over HasField is for records defined with data instance.

Methods

field_ :: Lens s t a b

getField :: forall f a s. HasField' f s a => s -> a

>>> getField @"age" human
50

setField :: forall f s a. HasField' f s a => a -> s -> s

>>> setField @"age" 60 human
Human {name = "Tunyasz", age = 60, address = "London", other = False}