Data.Generics.Sum

Description

Magic sum operations using Generics

These classes need not be instantiated manually, as GHC can automatically prove valid instances via Generics. Only the Generic class needs to be derived (see examples).

Prisms

module Data.Generics.Sum.Any

module Data.Generics.Sum.Constructors

module Data.Generics.Sum.Subtype

module Data.Generics.Sum.Typed